# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SkillTwin AI Mentor API",
    "description": "Contract and intelligence source-of-truth for the SkillTwin personalized AI mentor platform. Exposes adaptive goals, winding journey roadmaps, evidence-based learner twin state, real-time mentor guidance, and spaced retention retrieval practice.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/profile": {
      "get": {
        "tags": [
          "Profile"
        ],
        "summary": "Get Learner Profile",
        "description": "Returns the authenticated user's profile.\nExtracts identity strictly from verified Supabase JWT claims, ensuring no spoofing.",
        "operationId": "get_profile_api_v1_profile_get",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Profile"
        ],
        "summary": "Update Learner Profile",
        "description": "Updates the authenticated user's profile details.",
        "operationId": "update_profile_api_v1_profile_patch",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProfileUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/revision/next": {
      "get": {
        "tags": [
          "Revision"
        ],
        "summary": "Get Next Retrieval Practice Queue",
        "description": "Returns prioritized 3-10 minute retrieval practice items based on retention decay,\nmisconceptions, and spaced repetition intervals.",
        "operationId": "get_next_revision_api_v1_revision_next_get",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionNextResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/revision/{id}/complete": {
      "post": {
        "tags": [
          "Revision"
        ],
        "summary": "Complete Spaced Review",
        "description": "Completes a spaced retrieval review session.\nApplies RetentionEngine:\n- Gradually reduces review frequency (expands interval) on successful retrieval.\n- Contracts interval on failure.\n- Recalibrates next_review, retention_score, and composite priority.",
        "operationId": "complete_spaced_review_api_v1_revision__id__complete_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCompleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/revision/submit": {
      "post": {
        "tags": [
          "Revision"
        ],
        "summary": "Submit Retrieval Result",
        "description": "Submits spaced retrieval practice result and reschedules interval (backwards compatible).",
        "operationId": "submit_retrieval_review_api_v1_revision_submit_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewSubmissionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/revision/notifications": {
      "get": {
        "tags": [
          "Revision"
        ],
        "summary": "Get Mentor Revision Notifications",
        "description": "Returns active mentor notifications reminding the learner of decaying retention cards.",
        "operationId": "get_revision_notifications_api_v1_revision_notifications_get",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MentorNotificationsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resources/upload": {
      "post": {
        "tags": [
          "Resources"
        ],
        "summary": "Upload Resource (Async Pipeline)",
        "operationId": "upload_resource_api_v1_resources_upload_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_resource_api_v1_resources_upload_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resources": {
      "get": {
        "tags": [
          "Resources"
        ],
        "summary": "List Resources",
        "operationId": "list_resources_api_v1_resources_get",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResourceResponse"
                  },
                  "title": "Response List Resources Api V1 Resources Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resources/{resource_id}": {
      "get": {
        "tags": [
          "Resources"
        ],
        "summary": "Get Resource Details",
        "operationId": "get_resource_api_v1_resources__resource_id__get",
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resource Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Resources"
        ],
        "summary": "Delete Resource",
        "operationId": "delete_resource_api_v1_resources__resource_id__delete",
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resource Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resources/search": {
      "post": {
        "tags": [
          "Resources"
        ],
        "summary": "RAG Vector Search",
        "operationId": "search_resources_api_v1_resources_search_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RAGSearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RAGSearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_resource_api_v1_resources_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "is_public": {
            "type": "boolean",
            "title": "Is Public",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_resource_api_v1_resources_upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MentorNotificationsResponse": {
        "properties": {
          "unread_count": {
            "type": "integer",
            "title": "Unread Count"
          },
          "notifications": {
            "items": {
              "$ref": "#/components/schemas/MentorNotification"
            },
            "type": "array",
            "title": "Notifications"
          }
        },
        "type": "object",
        "required": [
          "unread_count",
          "notifications"
        ],
        "title": "MentorNotificationsResponse"
      },
...
      "RAGSearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "concept_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concept Id"
          },
          "top_k": {
            "type": "integer",
            "title": "Top K",
            "default": 4
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "RAGSearchRequest"
      },
...
      "ReviewCompleteRequest": {
        "properties": {
          "is_successful": {
            "type": "boolean",
            "title": "Is Successful",
            "description": "True if retrieval was correct/successful, False if failed"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence",
            "description": "Self-reported confidence percentage"
          },
          "time_spent_seconds": {
            "type": "integer",
            "title": "Time Spent Seconds",
            "description": "Time spent on retrieval practice",
            "default": 120
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "is_successful"
        ],
        "title": "ReviewCompleteRequest"
      },
      "ReviewItemResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "concept_id": {
            "type": "string",
            "title": "Concept Id"
          },
          "concept_name": {
            "type": "string",
            "title": "Concept Name"
          },
          "last_reviewed": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Reviewed"
          },
          "next_review": {
            "type": "string",
            "format": "date-time",
            "title": "Next Review"
          },
          "interval": {
            "type": "integer",
            "title": "Interval",
            "default": 1
          },
          "successful_retrievals": {
            "type": "integer",
            "title": "Successful Retrievals",
            "default": 0
          },
          "failed_retrievals": {
            "type": "integer",
            "title": "Failed Retrievals",
            "default": 0
          },
          "retention_score": {
            "type": "number",
            "title": "Retention Score",
            "default": 100.0
          },
          "priority_score": {
            "type": "number",
            "title": "Priority Score",
            "default": 0.0
          },
          "is_high_priority": {
            "type": "boolean",
            "title": "Is High Priority",
            "default": false
          },
          "retention_risk": {
            "type": "string",
            "title": "Retention Risk",
            "default": "medium"
          },
          "why_today": {
            "type": "string",
            "title": "Why Today",
            "default": "Memory decay threshold reached. Spaced retrieval anchors foundational invariants."
          },
          "mentor_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mentor Prompt"
          },
          "estimated_minutes": {
            "type": "integer",
            "title": "Estimated Minutes",
            "default": 4
          },
          "next_review_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Review Text"
          },
          "due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "interval_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interval Days"
          },
          "repetitions": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repetitions"
          },
          "last_reviewed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Reviewed At"
          },
          "ease_factor": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ease Factor",
            "default": 2.5
          }
        },
        "type": "object",
        "required": [
          "id",
          "concept_id",
          "concept_name",
          "next_review"
        ],
        "title": "ReviewItemResponse"
      },
      "ReviewSubmissionRequest": {
        "properties": {
          "review_item_id": {
            "type": "string",
            "title": "Review Item Id"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RetrievalResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "accuracy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RetrievalAccuracy"
              },
              {
                "type": "null"
              }
            ]
          },
          "confidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RetrievalConfidence"
              },
              {
                "type": "null"
              }
            ]
          },
          "time_spent_seconds": {
            "type": "integer",
            "title": "Time Spent Seconds",
            "default": 120
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "review_item_id"
        ],
        "title": "ReviewSubmissionRequest"
      },
      "RevisionNextResponse": {
        "properties": {
          "header": {
            "type": "string",
            "title": "Header",
            "default": "5 minutes for your future self."
          },
          "total_due_count": {
            "type": "integer",
            "title": "Total Due Count"
          },
          "primary_review_item": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewItemResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "upcoming_items": {
            "items": {
              "$ref": "#/components/schemas/ReviewItemResponse"
            },
            "type": "array",
            "title": "Upcoming Items",
            "default": []
          },
          "mentor_guidance": {
            "type": "string",
            "title": "Mentor Guidance",
            "default": "Consistent 5-minute retrieval is the single highest-yield cognitive investment."
          }
        },
        "type": "object",
        "required": [
          "total_due_count"
        ],
        "title": "RevisionNextResponse"
      },
...
      "UserProfileResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "default": "UTC"
          },
          "preferences": {
            "additionalProperties": true,
            "type": "object",
            "title": "Preferences"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "full_name",
          "created_at",
          "updated_at"
        ],
        "title": "UserProfileResponse"
      },
      "UserProfileUpdateRequest": {
        "properties": {
          "full_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          },
          "preferences": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferences"
          }
        },
        "type": "object",
        "title": "UserProfileUpdateRequest"
      },
...
    }
  }
}